Restructured card layout based on design changes.
